+++ R/get_estimates.R
@@ -10,7 +10,7 @@
 #' @param model a single model object
 #' 
 #' @export
-get_estimates <- function(model, conf_level = .95, vcov = NULL, shape = NULL, ...) {
+get_estimates <- function(model, conf_level = .95, vcov = NULL, shape = NULL, coef_rename = FALSE, ...) {
 
     if (is.null(conf_level)) {
         conf_int <- FALSE
@@ -52,6 +52,7 @@ get_estimates <- function(model, conf_level = .95, vcov = NULL, shape = NULL, ..
                 conf_int = conf_int,
                 conf_level = conf_level,
                 vcov = V,
+                coef_rename = coef_rename,
                 ...)
             if (is.character(out)) {
                 warning_msg <- c(warning_msg, out)
@@ -154,6 +155,7 @@ get_estimates_parameters <- function(model,
                                      conf_int,
                                      conf_level,
                                      vcov,
+                                     coef_rename,
                                      ...) {
 
     dots <- list(...)
@@ -180,12 +182,14 @@ get_estimates_parameters <- function(model,
     if (isTRUE(mi[["is_bayesian"]])) {
         if (!"test" %in% names(dots)) args <- c(args, list("test" = NULL))
         if (!"diagnostic" %in% names(dots)) args <- c(args, list("diagnostic" = NULL))
+        if (!"dispersion" %in% names(dots)) args <- c(args, list("dispersion" = TRUE))
     }
 
     # main call
     tidy_easystats <- function(...) {
         dots <- list(...)
         # bug in `parameters`
+        dots[["pretty_names"]] <- "labels"
         inner <- parameters::parameters
         out <- do.call("inner", dots)
         out <- parameters::standardize_names(out, style = "broom")
@@ -196,6 +200,14 @@ get_estimates_parameters <- function(model,
       args[["vcov"]] <- vcov
     }
     out <- hush(tryCatch(do.call("tidy_easystats", args), error = function(e) NULL))
+
+    if (isTRUE(coef_rename)) {
+        labs <- attr(out, "pretty_labels")
+        labs <- gsub("\\*", "\u00d7", labs)
+        if (isTRUE(length(labs) == nrow(out))) {
+            out$term <- labs
+        }
+    }
 
     # errors and warnings: before processing the data frame term names
     if (!inherits(out, "data.frame") || nrow(out) < 1) {