+++ tests/testthat/test-datasummary_balance.R
@@ -327,4 +327,3 @@ test_that('datasummary_balance: various datasets', {
 })
 
 
-