+++ tests/testthat/test-labels.R
@@ -1,7 +1,8 @@
+skip_if_not_installed("parameters", minimum_version = "0.18.2.9")
+
 requiet("haven")
 
 test_that("datasummary labels",{
-
     # datasummary()
     dat <- mtcars
     dat$am <- haven::labelled(
@@ -45,11 +46,11 @@ test_that("datasummary labels",{
     expect_true("Cylinders" %in% colnames(tab))
     expect_true("Transmission" %in% colnames(tab))
 
-    # datasummary_balance is not supported yet
-    expect_warning(expect_warning(
-        tab <- datasummary_balance(~am, data = dat),
-        regexp = "belled.*balance"),
-        regexp = "It is not safe")
+    # # datasummary_balance is not supported yet
+    # expect_warning(expect_warning(
+    #     tab <- datasummary_balance(~am, data = dat),
+    #     regexp = "belled.*balance"),
+    #     regexp = "It is not safe")
 })
 
 
@@ -70,7 +71,7 @@ test_that("modelsummary: use variable labels by default", {
   mod <- list(
     "Bivariate" = lm(Girth ~ Height, data = dat),
     "Multivariate" = lm(Girth ~ Height + Volume, data = dat))
-  tab <- modelsummary(mod, "dataframe")
+  tab <- modelsummary(mod, "dataframe", coef_rename = TRUE)
   expect_true(all(c("Height (in feet)", "Volume (in liters)") %in% tab$term))
 })
 
@@ -81,7 +82,7 @@ test_that("interaction", {
   dat$Height <- haven::labelled(dat$Height, label = "Height (in feet)")
   dat$Volume <- haven::labelled(dat$Volume, label = "Volume (in liters)")
   mod <- lm(Girth ~ Height*Volume, data = dat)
-  tab <- modelsummary(mod, "dataframe")
+  tab <- modelsummary(mod, output = "dataframe", coef_rename = TRUE)
   expect_true("Height (in feet) × Volume (in liters)" %in% tab$term)
 })
 
@@ -107,7 +108,7 @@ test_that("modelsummary: also applies variable labels for depvar", {
   mod <- list(
     lm(mpg ~ cyl + drat + disp, data = dat),
     lm(hp ~ cyl + drat + disp, data = dat))
-  tab <- modelsummary(dvnames(mod), "dataframe")
+  tab <- modelsummary(dvnames(mod), "dataframe", coef_rename = TRUE)
   expect_equal(names(tab)[4:5], c("Miles per gallon", "hp"))
   expect_true("Number of cylinders" %in% tab$term)
 })